Прикрутил абстрактный класс и немного прибрался в утилзах
